feat(identity): add isMeta property to scope and role models

+++ packages/identity/src/controllers/role.controller.ts
@@ -3,6 +3,8 @@ import { user } from "../resolvers/user.resolver";
 import { deleteNested } from "../utils/deleteNested";
 import { scope } from "../resolvers/scope.resolver";
 import { ScopeNotFoundError } from "../errors/ScopeNotFound.error";
+import { RoleIsMetaRoleError } from "../errors/RoleIsMetaRole.error";
+import { RoleNotFoundError } from "../errors/RoleNotFound.error";
 
 class RoleController extends Controller {
   getAllUsers = async (parent: any) =>
@@ -16,8 +18,14 @@ class RoleController extends Controller {
 
   async create(params: any) {
     const role = await super.create(params);
-    const { id: roleScopeId } = await scope.create({ name: role.id });
-    const { id: writeSelfRoleId } = await super.create({ name: "WRITE:ROLE" });
+    const { id: roleScopeId } = await scope.create({
+      name: role.id,
+      isMeta: true
+    });
+    const { id: writeSelfRoleId } = await super.create({
+      name: "WRITE:ROLE",
+      isMeta: true
+    });
 
     await scope.assignRole({
       scopeId: roleScopeId,
@@ -36,21 +44,25 @@ class RoleController extends Controller {
     const { id: roleId } = params;
     const role = await this.getWithScopes(roleId);
     if (!role) {
-      throw new ScopeNotFoundError();
+      throw new RoleNotFoundError();
     } else {
-      const { id: scopeId } = role.scopes.find(
-        (scope: any) => scope.name === roleId
-      );
+      if (role.isMeta) {
+        throw new RoleIsMetaRoleError();
+      } else {
+        const { id: scopeId } = role.scopes.find(
+          (scope: any) => scope.name === roleId
+        );
 
-      await user.auth({
-        ...params,
-        scopeId,
-        validRolesNames: ["WRITE:ROLE"]
-      });
+        await user.auth({
+          ...params,
+          scopeId,
+          validRolesNames: ["WRITE:ROLE"]
+        });
 
-      return super.update(roleId, {
-        name: params.name
-      });
+        return super.update(roleId, {
+          name: params.name
+        });
+      }
     }
   }
 
@@ -58,32 +70,36 @@ class RoleController extends Controller {
     const { id: roleId } = params;
     const role = await this.getWithScopes(roleId);
     if (!role) {
-      throw new ScopeNotFoundError();
+      throw new RoleNotFoundError();
     } else {
-      const { id: scopeId } = role.scopes.find(
-        (scope: any) => scope.name === roleId
-      );
+      if (role.isMeta) {
+        throw new RoleIsMetaRoleError();
+      } else {
+        const { id: scopeId } = role.scopes.find(
+          (scope: any) => scope.name === roleId
+        );
 
-      await user.auth({
-        ...params,
-        scopeId,
-        validRolesNames: ["WRITE:ROLE"]
-      });
+        await user.auth({
+          ...params,
+          scopeId,
+          validRolesNames: ["WRITE:ROLE"]
+        });
 
-      const deletedRole: any = await deleteNested(
-        this,
-        roleId,
-        async (deletableId: string) => await super.delete(deletableId),
-        [
-          {
-            getter: async (roleId: string) => (await this.get(roleId)).users,
-            controller: user,
-            path: "scopes"
-          }
-        ]
-      );
+        const deletedRole: any = await deleteNested(
+          this,
+          roleId,
+          async (deletableId: string) => await super.delete(deletableId),
+          [
+            {
+              getter: async (roleId: string) => (await this.get(roleId)).users,
+              controller: user,
+              path: "scopes"
+            }
+          ]
+        );
 
-      return deletedRole;
+        return deletedRole;
+      }
     }
   }
 }
